requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.4.post0
    # via discord.py
appdirs==1.4.4
    # via
    #   black
    #   urlextract
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
black==20.8b1
    # via -r requirements.in
chardet==4.0.0
    # via aiohttp
click==7.1.2
    # via black
discord.py==1.6.0
    # via discord
discord==1.0.1
    # via -r requirements.in
filelock==3.0.12
    # via urlextract
idna==3.1
    # via
    #   urlextract
    #   yarl
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
pathspec==0.8.1
    # via black
python-dotenv==0.15.0
    # via -r requirements.in
regex==2020.11.13
    # via black
toml==0.10.2
    # via black
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   black
    #   yarl
uritools==3.0.0
    # via urlextract
urlextract==1.2.0
    # via -r requirements.in
yarl==1.6.3
    # via aiohttp
